refactor(rust): move CP01 PyO3 binding into the rules crate
The cp01_violations pyfunction was sitting in the root crate's pymodule glue
(src/python.rs) — rule-specific code in generic registration wiring, which would
accrete a per-rule function for every future rule. Move it next to its logic:

- sqlfluffrs_rules gains an optional `python` feature (pyo3 + parser/python) and
  a `python` module holding the cp01_violations binding + a `register(m)` fn,
  mirroring how the lexer/parser crates own their Py* surfaces.
- The root crate's python.rs drops the inline function and just calls
  sqlfluffrs_rules::python::register(m); its `python` feature enables
  sqlfluffrs_rules/python.

Rule detection stays pure Rust (no feature needed); only the thin binding is
feature-gated. Python API surface (sqlfluffrs.cp01_violations(tree, …)) is
unchanged. Build + 38-case parity test green; cargo fmt/clippy clean.
