+++ sqlfluffrs/Cargo.toml
@@ -35,6 +35,7 @@ python = [
     "pyo3",
     "sqlfluffrs_lexer/python",
     "sqlfluffrs_parser/python",
+    "sqlfluffrs_rules/python",
     "sqlfluffrs_python",
 ]
