+++ sqlfluffrs/sqlfluffrs_rules/Cargo.toml
@@ -4,7 +4,19 @@ version.workspace = true
 edition = "2021"
 
 # Rust-native lint rules. Detection runs over the parse arena's public read API
-# (sqlfluffrs_parser::Arena); fixes are still applied Python-side for now. No
-# PyO3 here — the thin bindings live in the root crate, which depends on both.
+# (sqlfluffrs_parser::Arena); fixes are still applied Python-side for now. The
+# rule logic is pure Rust; the optional `python` feature adds each rule's thin
+# PyO3 binding, registered by the root crate.
 [dependencies]
 sqlfluffrs_parser = { path = "../sqlfluffrs_parser" }
+# NOTE: Update the abi3 version as we deprecate old python versions.
+# It should match the oldest currently supported python version.
+pyo3 = { version = "0.29.0", optional = true, features = [
+    "abi3-py310",
+    "hashbrown",
+    "extension-module",
+    "uuid",
+] }
+
+[features]
+python = ["pyo3", "sqlfluffrs_parser/python"]
